requirements.txt
praw>=7.7.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0

